fix(devex): reconcile the branch with what master changed underneath it
Master moved two things this branch's rebase couldn't see.

Stamphog's PullRequest serializer no longer exposes audience_key or
digest_run — with several audiences per PR those live on the per-audience
rows now, and the facade DTO dropped them. This branch had converted the
same serializer to a DataclassSerializer over the contract, so the two
edits landed on top of each other: a declared digest_run field with no
entry in Meta.fields, which trips DRF's assertion the moment the
serializer is constructed.

ManagedWarehouseSQLMode is new on master and reaches core through
warehouse_sources' facade/models.py, which the isolation check added here
rejects — it's a StrEnum, not a Django model, so the watched-models
allowance doesn't cover it and shouldn't grow to. It belongs in
facade/types.py with the other shared enums, where the contract-check
inputs already watch it; backend/types.py re-exports it for
product-internal use like the rest.
